ツールチップの消費マナを表示しつつ、クールダウン表示をイベント駆動ではないように変更
