+++ src/main/java/jp/aquafactory/apprenticecodex/item/curios/autocastamulet/AutocastAmuletClientNotificationState.java
@@ -1,16 +1,8 @@
 package jp.aquafactory.apprenticecodex.item.curios.autocastamulet;
 
-import io.redspace.ironsspellbooks.api.spells.SpellData;
-import io.redspace.ironsspellbooks.player.ClientMagicData;
-import net.minecraft.ChatFormatting;
 import net.minecraft.client.Minecraft;
-import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.world.item.ItemStack;
 import org.jetbrains.annotations.Nullable;
-
-import java.util.ArrayList;
-import java.util.List;
 
 public final class AutocastAmuletClientNotificationState {
     private static final AutocastAmuletNotificationController CONTROLLER = new AutocastAmuletNotificationController();
@@ -55,47 +47,6 @@ public final class AutocastAmuletClientNotificationState {
         return CONTROLLER.getActiveNotification();
     }
 
-    public static List<Component> buildCooldownTooltipLines(ItemStack stack) {
-        var lines = new ArrayList<Component>();
-        var player = Minecraft.getInstance().player;
-        if (player == null) {
-            return lines;
-        }
-
-        var activeCooldowns = new ArrayList<SpellCooldownView>();
-        for (SpellData spellData : AutocastAmulet.getImbuedSpells(stack)) {
-            var spell = spellData.getSpell();
-            if (spell == null) {
-                continue;
-            }
-
-            var cooldown = ClientMagicData.getCooldowns().getSpellCooldowns().get(spell.getSpellId());
-            if (cooldown == null || cooldown.getCooldownRemaining() <= 0.0F) {
-                continue;
-            }
-
-            activeCooldowns.add(new SpellCooldownView(
-                    spell.getDisplayName(player),
-                    AutocastAmuletNotificationController.toDisplayCooldownSeconds(cooldown.getCooldownRemaining())
-            ));
-        }
-
-        if (activeCooldowns.isEmpty()) {
-            return lines;
-        }
-
-        lines.add(Component.empty());
-        lines.add(Component.translatable("item.apprenticecodex.autocast_amulet.cooldown_title").withStyle(ChatFormatting.GRAY));
-        for (var cooldownView : activeCooldowns) {
-            lines.add(Component.translatable(
-                    "item.apprenticecodex.autocast_amulet.cooldown_line",
-                    cooldownView.spellName(),
-                    cooldownView.remainingSeconds()
-            ).withStyle(ChatFormatting.DARK_AQUA));
-        }
-        return lines;
-    }
-
     public static float getActiveNotificationAlpha() {
         var gameTime = resolveCurrentGameTime();
         if (gameTime < 0L) {
@@ -112,8 +63,5 @@ public final class AutocastAmuletClientNotificationState {
         }
 
         return minecraft.level.getGameTime();
-    }
-
-    private record SpellCooldownView(Component spellName, int remainingSeconds) {
     }
 }
