+++ src/main/resources/assets/apprenticecodex/lang/en_us.json
@@ -135,8 +135,10 @@
   "item.apprenticecodex.autocast_amulet.desc_1": "Automatically casts imbued magic.",
   "item.apprenticecodex.autocast_amulet.desc_2": "Limited to certain spells, and will continuously attempt to cast.",
   "item.apprenticecodex.autocast_amulet.tooltip.hint": "Hold Shift for spell details.",
-  "item.apprenticecodex.autocast_amulet.cooldown_title": "Current spell cooldowns:",
-  "item.apprenticecodex.autocast_amulet.cooldown_line": "- %1$s: %2$ds",
+  "item.apprenticecodex.autocast_amulet.tooltip.status_title": "Autocast status:",
+  "item.apprenticecodex.autocast_amulet.tooltip.mana_multiplier": "Current mana cost multiplier: %1$s",
+  "item.apprenticecodex.autocast_amulet.tooltip.ready_line": "- %1$s: Ready (Mana Cost %2$d)",
+  "item.apprenticecodex.autocast_amulet.tooltip.cooldown_line": "- %1$s: Cooling down (%2$ds)",
   "item.apprenticecodex.mana_shield_charm.desc_1": "Absorbs damage by consuming mana.",
   "item.apprenticecodex.mana_shield_charm.desc_2": "Mana barrier does not benefit from armor.",
   "item.apprenticecodex.mana_shield_charm.desc_3": "Stops functioning until mana is fully recovered after depletion.",
