+++ src/main/resources/assets/apprenticecodex/lang/ja_jp.json
@@ -135,8 +135,10 @@
   "item.apprenticecodex.autocast_amulet.desc_1": "注入した魔法を自動で詠唱する。",
   "item.apprenticecodex.autocast_amulet.desc_2": "注入できる魔法に制限がある上に、常に詠唱し続けようとする。",
   "item.apprenticecodex.autocast_amulet.tooltip.hint": "Shiftキーで魔法の詳細を表示",
-  "item.apprenticecodex.autocast_amulet.cooldown_title": "現在クールダウン中の魔法:",
-  "item.apprenticecodex.autocast_amulet.cooldown_line": "- %1$s: %2$d秒",
+  "item.apprenticecodex.autocast_amulet.tooltip.status_title": "自動詠唱の状態:",
+  "item.apprenticecodex.autocast_amulet.tooltip.mana_multiplier": "現在の消費マナ倍率: %1$s",
+  "item.apprenticecodex.autocast_amulet.tooltip.ready_line": "- %1$s: 使用可能(消費マナ%2$d)",
+  "item.apprenticecodex.autocast_amulet.tooltip.cooldown_line": "- %1$s: クールダウン中(%2$d秒)",
   "item.apprenticecodex.mana_shield_charm.desc_1": "マナを消費してダメージを打ち消す。",
   "item.apprenticecodex.mana_shield_charm.desc_2": "ただし、マナの障壁は防御力の効果を受けることができない。",
   "item.apprenticecodex.mana_shield_charm.desc_3": "マナを使い果たすとマナが全回復するまで機能を停止する。",
